+++ src/Entity/DocumentEntityCollection.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+
+namespace XDevApi\Entity;
+
+
+use Zend\Paginator\Paginator;
+
+final class DocumentEntityCollection extends Paginator
+{
+}